Added README and small logic tweak
Added check to order logic which prevents submitted orders from being deleted.

+++ src/main/java/com/miloske/tokyotask/controller/OrderController.java
@@ -110,6 +110,10 @@ public class OrderController {
 				return new ResponseEntity<String>(HttpStatus.UNAUTHORIZED);
 			}
 			
+			if(order.getOrderStatus() == OrderStatus.SUBMITTED ) {
+				return new ResponseEntity<String>(HttpStatus.FORBIDDEN);
+			}
+			
 			orderRepository.delete(order);
 			return new ResponseEntity<String>(String.format("Order %d deleted", id), HttpStatus.OK);
 		}).orElseGet(() -> {
